travaia-ai/backend/ai-engine-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-aiplatform
google-cloud-texttospeech==2.14.2
google-cloud-speech==2.21.0
vertexai
pydantic==2.5.0
structlog==23.2.0
httpx==0.25.2
slowapi==0.1.9
google-cloud-pubsub==2.18.4
firebase-admin==6.2.0
pydantic[email]==2.5.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
tenacity==8.2.3
python-jose[cryptography]==3.3.0